[MVP-5] Temp stop point.
